# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BugShot API Documentation",
        "description": "OpenApi documentation using swagger for BugShot project.",
        "contact": {
            "email": "info@it-michel.de"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.test/api/v1"
        }
    ],
    "paths": {
...
        "/bugs/{bug_id}/assign-user": {
            "post": {
                "tags": [
                    "Bug"
                ],
                "summary": "Assign user to bug.",
                "operationId": "assignUser",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "bug_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Bug/properties/id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "The id of the user",
                                        "type": "integer",
                                        "format": "int64"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/bugs/{bug_id}/users": {
            "get": {
                "tags": [
                    "Bug"
                ],
                "summary": "All bug users.",
                "operationId": "allBugUsers",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "name": "bug_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Bug/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BugUserRole"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/bugs/{bug_id}/users/{user_id}": {
            "delete": {
                "tags": [
                    "Bug"
                ],
                "summary": "Remove user from the bug.",
                "operationId": "removeBugUser",
                "parameters": [
                    {
                        "name": "bug_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Bug/properties/id"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/User/properties/id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BugUserRole": {
                "properties": {
                    "bug_id": {
                        "description": "The id of the bug.",
                        "type": "string",
                        "maxLength": 255
                    },
                    "user_id": {
                        "description": "The id of the user.",
                        "type": "integer",
                        "format": "int64"
                    },
                    "role_id": {
                        "description": "The id of the role.",
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        []
    ]
}